backend/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask==3.1.1

mysql-connector-python=9.3.0